Define a Data Transformation Schematic

define() allows one to operate on data using one or more formula-based definitions. Data transformation and selection can be achieved with formulas or using standard data.table syntax in a procedural manner with a single function call. Each operation operates on the previous result making define() a pipeline of operations in a compact framework.

Usage

define(
  data = NULL,
  ...,
  keep.rownames = TRUE,
  blueprint = NULL,
  progress = FALSE
)

Arguments

data

The input data (see 'Details')

...

(dots_list): Formulas for which the left-hand side (LHS) is an expression containing the operation, and the right-hand side (RHS) contains column names that form a grouping set for the operation (i.e., <expression> ~ col_1 + col_2 + ...):

If the form <LHS>~ . is given, the LHS executes using all columns as the grouping set

If the form <LHS>~ 1 is given, the LHS executes without grouping

If no LHS is given, the operation defaults to selection based on the RHS

keep.rownames

See data.table

blueprint

(EXPERIMENTAL) See blueprint. Blueprints are always evaluated first when provided: this makes it easy to "branch" off of a base dataset.

progress

(logical | FALSE) Should a progress bar be shown during execution (passed to map)?

Value

The data modified

Details

If data is a smart.data object, the taxonomical references to field names can be accessed by using the use() syntax in the right-hand side of the formula (e.g., ~use(term1, term2) + otherTerm1 + ...).

When using formulas for ... on an empty data.table object, set the argument data to NULL; otherwise, you will get an error.