Create a Data Join Map

join.mapper Creates a map with which datasets can be joined using data.table methods. Cross-environment objects are not supported, but objects in attached environment might work. Also, the output is given attribute "env" to store the value of env

join.mapper(
  map_name = "new_join_map",
  env = parent.frame(),
  obj_names,
  field_names = "*",
  clean = FALSE
)

Arguments

map_name

The output map name to use in assignment

env

The environment where source objects are found; this is also the environment of the assigned output.

obj_names

(string[]) The names of objects to join

field_names

(string[]) One or more strings and REGEX patterns used to define the field names to use for possible joins: matching is either REGEX or identity, and support for aliasing via "primary_col==alias_col" (note the quotes) is supported.

clean

(logical) TRUE indicates that the object should be removed from env before creating the output

Value

If a DBOE DSN environment is passed to env, a local environment is used to contain the output of special code that transforms env$metamap into environment objects that can be used for creating the map: this should be stored or piped into join.reduce; otherwise, a data.table object used join datasets via join.reduce