Join-Reduce Multiple Datasets

join.reduce leverages data.table fast joins and reduce from package purr

join.reduce(
  join_map,
  out_name,
  x_names,
  i_names,
  filters = TRUE,
  dt_key,
  source_env = attr(join_map, "env"),
  assign_env = attr(jmap, "env"),
  clean = FALSE
)

Arguments

join_map

The input join-map (see link{join.mapper})

out_name

The name of the output object

x_names

(string[]) Names or REGEX patterns indicating the outer table: multiple values will be concatenated into a delimited string

i_names

(string[]) Names or REGEX patterns indicating the inner table to be joined: multiple values will be concatenated into a delimited string

filters

(expression[[]]) data.table-friendly expression to limit the rows of the output. A length-2 list indicates row and column expressions respectively.

dt_key

(string[], symbol[]) Names that participate in creating the key for the output

source_env

The environment from which the input should be sourced. If empty, it defaults to attr(jmap, "env")

assign_env

The environment in which the output should be assigned. If empty, it defaults to attr(jmap, "env")

clean

(logical) TRUE indicates that the object should be removed from env before creating the output

Value

Combined datasets assigned to the designated environment