Define a Data Transformation Schematic

define() allows one to operate on data using one or more formula-based definitions. Data transformation and selection can be achieved with formulas or using standard data.table syntax in a procedural manner with a single function call. Each operation operates on the previous result making define() a pipeline of operations in a compact framework.

define(data = NULL, ..., keep.rownames = TRUE, blueprint = NULL)

Arguments

data

The input data (see 'Details')

...

(dots_list): Formulas for which the left-hand side (LHS) is an expression containing the operation, and the right-hand side (RHS) contains column names that form a grouping set for the operation (i.e., <expression> ~ col_1 + col_2 + ...):

If the form <LHS>~ . is given, the LHS executes using all columns as the grouping set

If the form <LHS>~ 1 is given, the LHS executes without grouping

If no LHS is given, the operation defaults to selection based on the RHS

keep.rownames

See data.table

blueprint

(EXPERIMENTAL) See blueprint.

Value

The data modified

Details

If data is a smart.data object, the taxonomical references to field names can be accessed by using the use() syntax in the right-hand side of the formula (e.g., ~use(term1, term2) + otherTerm1 + ...).

When using formulas for ... on an empty data.table object, set the argument data to NULL; otherwise, you will get an error.

Examples

library(smart.data)
#> Loading required package: magrittr
smart.start();

taxonomy_list <- list(
  identifier = taxonomy(
    term = "identifier"
    , desc = "Identifier"
    , fields = c("rn"))
  , category = taxonomy(
    term = "category"
    , desc = "Category"
    , fields = c("cyl", "gear", "carb")
    )
);

smart_mt <- smart.data$
  new(as.data.table(mtcars[1:10, ], keep.rownames = TRUE))$
  taxonomy.rule(!!!taxonomy_list)$
  enforce.rules(for_usage)$
  cache_mgr(action = upd);
#> 'new_data' is not found in the smart-cache

define(
  smart_mt
  , list(j = 1, mpg) ~vs + am + use(identifier, category)
  , ~j + mpg
  )[];
#>         j   mpg
#>     <num> <num>
#>  1:     1  21.0
#>  2:     1  21.0
#>  3:     1  22.8
#>  4:     1  21.4
#>  5:     1  18.7
#>  6:     1  18.1
#>  7:     1  14.3
#>  8:     1  24.4
#>  9:     1  22.8
#> 10:     1  19.2

define(smart_mt, ~vs + am + use(identifier, category))[];
#>        vs    am                rn   cyl  gear  carb
#>     <num> <num>            <char> <num> <num> <num>
#>  1:     0     1         Mazda RX4     6     4     4
#>  2:     0     1     Mazda RX4 Wag     6     4     4
#>  3:     1     1        Datsun 710     4     4     1
#>  4:     1     0    Hornet 4 Drive     6     3     1
#>  5:     0     0 Hornet Sportabout     8     3     2
#>  6:     1     0           Valiant     6     3     1
#>  7:     0     0        Duster 360     8     3     4
#>  8:     1     0         Merc 240D     4     4     2
#>  9:     1     0          Merc 230     4     4     2
#> 10:     1     0          Merc 280     6     4     4

define(smart_mt, x = sum(am^2) ~ use(identifier, category))[];
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").
define()[];
#> Null data.table (0 rows and 0 cols)
define(x = 1:10, y = x * 3)[];
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").
define(x = 1:10, y = x * 3, z = x*y)[];
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").
define(NULL, x = 1:10, y = x * 3, z = x*y, ~x + z)[];
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").
define(data.table(), x = 1:10, y = x * 3, list(z = 10) ~ x)[];
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").

# Predefined operations:
predef_data <- define(smart_mt, x = sum(am^2) ~ use(identifier, category));
#> Error in map2(.x, .y, .f, ..., .progress = .progress): ℹ In index: 1.
#> ℹ With name: x.
#> Caused by error:
#> ! Check that is.data.table(DT) == TRUE. Otherwise, :=, `:=`(...) and let(...) are defined for use in j, once only and in particular ways. See help(":=").

redef_data <- define(
  smart_mt
  # Normally, listing 'x' would throw an error as it would not exist in the data;
  # however, since a blueprint is provided from a previous definition, the 'x' variable is
  # created within scope before the additional operations execute:
  , list(j = 1, mpg, x) ~vs + am + use(identifier, category)
  , blueprint = predef_data
  );
#> Error in eval(expr, envir, enclos): object 'predef_data' not found

redef_data;
#> Error in eval(expr, envir, enclos): object 'redef_data' not found

attr(redef_data, "blueprint");
#> Error in eval(expr, envir, enclos): object 'redef_data' not found

identical(redef_data, define(smart_mt, blueprint = redef_data));
#> Error in eval(expr, envir, enclos): object 'redef_data' not found