Define a Data Operation

define allows one to operate on data using one or more formula-based definitions. Data transformation and selection can be achieved with formulas or using standard data.table syntax in a procedural manner with a single function call.

define(data = NULL, ..., keep.rownames = TRUE)

Arguments

data

The input data (see 'Details')

...

(dots_list): Formulas for which the left-hand side (LHS) is an expression containing the operation, and the right-hand side (RHS) contains column names that form a grouping set for the operation (i.e., <expression> ~ col_1 + col_2 + ...):

If the form <LHS>~ . is given, the LHS executes using all columns as the grouping set

If the form <LHS>~ 1 is given, the LHS executes without grouping

If no LHS is given, the operation defaults to selection based on the RHS

keep.rownames

See data.table

Value

The data modified

Details

If data is a smart.data object, the taxonomical references to field names can be accessed by using the use() syntax in the right-hand side of the formula (e.g., ~use(term1, term2) + otherTerm1 + ...).

When using formulas for ... on an empty data.table object, set the argument data to NULL; otherwise, you will get an error.

Examples

library(smart.data)
#> Loading required package: magrittr
smart.start();

taxonomy_list <- list(
  identifier = taxonomy(
    term = "identifier"
    , desc = "Identifier"
    , fields = c("rn"))
  , category = taxonomy(
    term = "category"
    , desc = "Category"
    , fields = c("cyl", "gear", "carb")
    )
);

smart_mt <- smart.data$
  new(as.data.table(mtcars, keep.rownames = TRUE))$
  taxonomy.rule(!!!taxonomy_list)$
  enforce.rules(for_usage)$
  cache_mgr(action = upd);
#> 'new_data' is not found in the smart-cache

define(
  smart_mt
  , list(j = 1, mpg) ~vs + am + use(identifier, category)
  , ~j + mpg
  )[];
#>     j  mpg
#>  1: 1 21.0
#>  2: 1 21.0
#>  3: 1 22.8
#>  4: 1 21.4
#>  5: 1 18.7
#>  6: 1 18.1
#>  7: 1 14.3
#>  8: 1 24.4
#>  9: 1 22.8
#> 10: 1 19.2
#> 11: 1 17.8
#> 12: 1 16.4
#> 13: 1 17.3
#> 14: 1 15.2
#> 15: 1 10.4
#> 16: 1 10.4
#> 17: 1 14.7
#> 18: 1 32.4
#> 19: 1 30.4
#> 20: 1 33.9
#> 21: 1 21.5
#> 22: 1 15.5
#> 23: 1 15.2
#> 24: 1 13.3
#> 25: 1 19.2
#> 26: 1 27.3
#> 27: 1 26.0
#> 28: 1 30.4
#> 29: 1 15.8
#> 30: 1 19.7
#> 31: 1 15.0
#> 32: 1 21.4
#>     j  mpg

define(smart_mt, ~vs + am + use(identifier, category))[];
#>     vs am                  rn cyl gear carb
#>  1:  0  1           Mazda RX4   6    4    4
#>  2:  0  1       Mazda RX4 Wag   6    4    4
#>  3:  1  1          Datsun 710   4    4    1
#>  4:  1  0      Hornet 4 Drive   6    3    1
#>  5:  0  0   Hornet Sportabout   8    3    2
#>  6:  1  0             Valiant   6    3    1
#>  7:  0  0          Duster 360   8    3    4
#>  8:  1  0           Merc 240D   4    4    2
#>  9:  1  0            Merc 230   4    4    2
#> 10:  1  0            Merc 280   6    4    4
#> 11:  1  0           Merc 280C   6    4    4
#> 12:  0  0          Merc 450SE   8    3    3
#> 13:  0  0          Merc 450SL   8    3    3
#> 14:  0  0         Merc 450SLC   8    3    3
#> 15:  0  0  Cadillac Fleetwood   8    3    4
#> 16:  0  0 Lincoln Continental   8    3    4
#> 17:  0  0   Chrysler Imperial   8    3    4
#> 18:  1  1            Fiat 128   4    4    1
#> 19:  1  1         Honda Civic   4    4    2
#> 20:  1  1      Toyota Corolla   4    4    1
#> 21:  1  0       Toyota Corona   4    3    1
#> 22:  0  0    Dodge Challenger   8    3    2
#> 23:  0  0         AMC Javelin   8    3    2
#> 24:  0  0          Camaro Z28   8    3    4
#> 25:  0  0    Pontiac Firebird   8    3    2
#> 26:  1  1           Fiat X1-9   4    4    1
#> 27:  0  1       Porsche 914-2   4    5    2
#> 28:  1  1        Lotus Europa   4    5    2
#> 29:  0  1      Ford Pantera L   8    5    4
#> 30:  0  1        Ferrari Dino   6    5    6
#> 31:  0  1       Maserati Bora   8    5    8
#> 32:  1  1          Volvo 142E   4    4    2
#>     vs am                  rn cyl gear carb

define(smart_mt, x = sum(am^2) ~ use(identifier, category))[];
#>                      rn  mpg cyl  disp  hp drat    wt  qsec vs am gear carb x
#>  1:           Mazda RX4 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4 1
#>  2:       Mazda RX4 Wag 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4 1
#>  3:          Datsun 710 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1 1
#>  4:      Hornet 4 Drive 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1 0
#>  5:   Hornet Sportabout 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2 0
#>  6:             Valiant 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1 0
#>  7:          Duster 360 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4 0
#>  8:           Merc 240D 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2 0
#>  9:            Merc 230 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2 0
#> 10:            Merc 280 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4 0
#> 11:           Merc 280C 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4 0
#> 12:          Merc 450SE 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3 0
#> 13:          Merc 450SL 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3 0
#> 14:         Merc 450SLC 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3 0
#> 15:  Cadillac Fleetwood 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4 0
#> 16: Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4 0
#> 17:   Chrysler Imperial 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4 0
#> 18:            Fiat 128 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1 1
#> 19:         Honda Civic 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2 1
#> 20:      Toyota Corolla 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1 1
#> 21:       Toyota Corona 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1 0
#> 22:    Dodge Challenger 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2 0
#> 23:         AMC Javelin 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2 0
#> 24:          Camaro Z28 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4 0
#> 25:    Pontiac Firebird 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2 0
#> 26:           Fiat X1-9 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1 1
#> 27:       Porsche 914-2 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2 1
#> 28:        Lotus Europa 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2 1
#> 29:      Ford Pantera L 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4 1
#> 30:        Ferrari Dino 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6 1
#> 31:       Maserati Bora 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8 1
#> 32:          Volvo 142E 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2 1
#>                      rn  mpg cyl  disp  hp drat    wt  qsec vs am gear carb x
define()[]
#> Null data.table (0 rows and 0 cols)
define(x = 1:10, y = x * 3)[]
#>      x  y
#>  1:  1  3
#>  2:  2  6
#>  3:  3  9
#>  4:  4 12
#>  5:  5 15
#>  6:  6 18
#>  7:  7 21
#>  8:  8 24
#>  9:  9 27
#> 10: 10 30
define(x = 1:10, y = x * 3, z = x*y)[]
#>      x  y   z
#>  1:  1  3   3
#>  2:  2  6  12
#>  3:  3  9  27
#>  4:  4 12  48
#>  5:  5 15  75
#>  6:  6 18 108
#>  7:  7 21 147
#>  8:  8 24 192
#>  9:  9 27 243
#> 10: 10 30 300
define(NULL, x = 1:10, y = x * 3, z = x*y, ~x + z)[]
#>      x   z
#>  1:  1   3
#>  2:  2  12
#>  3:  3  27
#>  4:  4  48
#>  5:  5  75
#>  6:  6 108
#>  7:  7 147
#>  8:  8 192
#>  9:  9 243
#> 10: 10 300
define(data.table(), x = 1:10, y = x * 3, list(z = 10) ~ x)[]
#>      x  z
#>  1:  1 10
#>  2:  2 10
#>  3:  3 10
#>  4:  4 10
#>  5:  5 10
#>  6:  6 10
#>  7:  7 10
#>  8:  8 10
#>  9:  9 10
#> 10: 10 10